backend/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2
joblib==1.3.2

# Deep Learning (for advanced features)
tensorflow==2.13.0
keras==2.13.1

# API Calls
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Chatbot
openai==0.27.8
google-cloud-dialogflow==2.23.0

# Voice Features
speechrecognition==3.10.0
pyttsx3==2.90
gtts==2.3.2

# Utilities
schedule==1.2.0
celery==5.3.1
redis==4.6.0

# Production
gunicorn==21.2.0